+++ packages/app/src/microsoft/teams/app/auth/__init__.py
@@ -3,13 +3,10 @@ Copyright (c) Microsoft Corporation. All rights reserved.
 Licensed under the MIT License.
 """
 
-from .jwt_middleware import JwtValidationMiddleware, create_jwt_validation_middleware
-from .service_token_validator import (
-    ServiceTokenValidator,
-)
+from .jwt_middleware import create_jwt_validation_middleware
+from .service_token_validator import ServiceTokenValidator
 
 __all__ = [
     "ServiceTokenValidator",
-    "JwtValidationMiddleware",
     "create_jwt_validation_middleware",
 ]
